Move 1D vertical gird functions to grid_1d module
